+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased changes
+* `Trigger` のハンドラ関数に非同期関数が渡せるように
+
 ## 1.0.1
 * `Trigger#addOnce()` のハンドラ内で `destroy()` が呼ばれた際にクラッシュしてしまうバグを修正